requirements.txt
djangorestframework==3.14.0
requests==2.28.2
Django~=4.1.5
pytest~=7.2.2
django_extensions==3.2.1
python-dotenv==1.0.0